Change PodPreset client in BUC from k8s to service-catalog one
