+++ components/binding-usage-controller/pkg/apis/settings/group.go
@@ -1,17 +1,6 @@
-/*
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
 // Package settings contains settings API versions
+// This is base on the types defined in this repository
+// https://github.com/jpeeler/podpreset-crd/tree/f726860d9e545790c6487291af6f77de8f3b5812/pkg/apis/settings
+// We need to define it here to be able to build the client. It's a temporary solution until the PodPreset from Service Catalog
+// will be finalized.
 package settings
